# src/endpoints/models/pvz.py
# coding: utf-8

"""
    backend service

    Сервис для управления ПВЗ и приемкой товаров

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, Field, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class PVZ(BaseModel):
    """
    PVZ
    """ # noqa: E501
    id: Optional[StrictStr] = None
    registration_date: Optional[datetime] = Field(default=None, alias="registrationDate")
    city: StrictStr
    __properties: ClassVar[List[str]] = ["id", "registrationDate", "city"]

    @field_validator('city')
    def city_validate_enum(cls, value):
        """Validates the enum"""
        if value not in ('Москва', 'Санкт-Петербург', 'Казань',):
            raise ValueError("must be one of enum values ('Москва', 'Санкт-Петербург', 'Казань')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        return self.model_dump_json(by_alias=True, exclude_unset=True)

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of PVZ from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of PVZ from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "registrationDate": obj.get("registrationDate"),
            "city": obj.get("city")
        })
        return _obj